# costa: install Node and Python packages in parallel