--- a/packages/costa/benchmark/makefile
+++ b/packages/costa/benchmark/makefile
@@ -2,3 +2,5 @@ export DEBUG=costa*
 
 bootstrap:
 	time node ./bootstrap.js
+install:
+	time node ./plugin-install.js