--- a/packages/costa/src/utils.ts
+++ b/packages/costa/src/utils.ts
@@ -1,6 +1,7 @@
 export interface LogStdio {
   stdout: NodeJS.WritableStream;
   stderr: NodeJS.WritableStream;
+  prefix?: string;
 }
 
 /**
@@ -8,13 +9,24 @@ export interface LogStdio {
  * Readable.pipi() will close the target pipe when end, we should ignore the end event.
  * @param readable child process stdout/stderr
  * @param writable the log stream
+ * @param prefix the log prefix
  */
-export function pipeLog(readable: NodeJS.ReadableStream, writable: NodeJS.WritableStream): void {
+export function pipeLog(readable: NodeJS.ReadableStream, writable: NodeJS.WritableStream, prefix?: string): void {
+  let buffer = '';
   readable.on('error', (err) => {
     writable.emit('error', err);
   });
-  readable.on('data', async (data) => {
-    writable.write(data);
+  readable.on('data', (data) => {
+    if (prefix) {
+      buffer += data.toString();
+      const list = buffer.split(/\n|\r/);
+      buffer = list.pop();
+      list.forEach((log) => {
+        writable.write(`${prefix}: ${log}\n`);
+      });
+    } else {
+      writable.write(data);
+    }
   });
 }