Benchmark:
Test command:
```shell
$ time ./packages/cli/dist/bin/pipcook plugin install @pipcook/plugins-tensorflow-resnet-model-define --tuna
```

This plugin depends on `tensorflow@2.2.0`.

| Branch                       | Test A          | Test B          | Test C          | Test D          | AVG             |
| ---------------------------- | --------------- | --------------- | --------------- | --------------- | --------------- |
| master                       | 1 m 20.06 s | 1 m 19.86 s | 1 m 25.03 s | 1 m 18.33 s | 1 m 20.82 s |
| modify/parallel-installation | 1 m 11.06 s | 1 m 04.47 s | 1 m 04.92 s | 1 m 11.27 s | 1 m 07.93 s |

In this typical scenario, reduced installation time by 15%.