@yorkie 2020-08-19
That would be that great if we add this benchmark script in this PR :)